Java/integ acarbo add xtrim refactor options (#253)
* Refactor stream options into a package


* Fix linking


* SPOTLESS


---------
